Mejor redesign to use proxy lambda
